Align github repository with Azure devops code.
The original place of the repository is in Azure devops repositories.
After this commit, the codebase is now transitioned to nordic SDK
v2.6.0.
Also a new project is added, central_wifi, which is developed on
nrf7002dk and combines the functionality of central_nbiot and
central_node and instead of nbiot, it uses wifi.

Signed-off-by: George Pappas <pappasgeorge12@gmail.com>